updated tooltips sample
