+++ 01-GettingStarted/client/index.html
@@ -15,7 +15,7 @@
     <script src="https://cdn.jsdelivr.net/npm/jquery@3.6.0/dist/jquery.min.js" ></script>
     <script src="https://cdn.quilljs.com/1.3.6/quill.min.js" ></script>    
     <script src="https://unpkg.com/dayjs@1.8.21/dayjs.min.js" ></script>    
-    <script src="https://dl.revealbi.io/reveal/libs/1.2.3/infragistics.reveal.js"></script>
+    <script src="https://dl.revealbi.io/reveal/libs/1.3.0/infragistics.reveal.js"></script>
 
     <script type="text/javascript">
         //set this to your server url
