+++ Tooltips/client/index.html
@@ -15,25 +15,24 @@
     <script src="https://cdn.jsdelivr.net/npm/jquery@3.6.0/dist/jquery.min.js"></script>
     <script src="https://cdn.quilljs.com/1.3.6/quill.min.js"></script>
     <script src="https://unpkg.com/dayjs@1.8.21/dayjs.min.js"></script>
-    <script src="https://dl.revealbi.io/reveal/libs/1.1.7/infragistics.reveal.js"></script>
+    <script src="https://dl.revealbi.io/reveal/libs/1.3.0/infragistics.reveal.js"></script>
 
     <script type="text/javascript">
 
         //set this to your server url
-        $.ig.RevealSdkSettings.setBaseUrl("https://localhost:7111/");
+        $.ig.RevealSdkSettings.setBaseUrl("http://localhost:5111/");
 
-        $.ig.RevealSdkSettings.ensureFontsLoadedAsync().then(() => {
-            $.ig.RVDashboard.loadDashboard("Sales", (dashboard) => {
-                var revealView = new $.ig.RevealView("#revealView");
-                revealView.dashboard = dashboard;
+        var revealView = new $.ig.RevealView("#revealView");
+        revealView.onTooltipShowing = (args) => {
+            if (args.visualization.title == "Sales") {
+                args.cancel = true;
+            }
+        };
 
-                revealView.onTooltipShowing = (args) => {
-                    if (args.visualization.title == "Sales") {
-                        args.cancel = true;
-                    }
-                };
-            });
-        });
+        $.ig.RVDashboard.loadDashboard("Sales").then(dashboard => {
+            revealView.dashboard = dashboard;
+        })
+        
     </script>
 </body>
