--- Tooltips/server/RevealSdk.Server/RevealSdk.Server.csproj
@@ -1,18 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <Nullable>enable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Reveal.Sdk.Web.AspNetCore.Trial" Version="1.1.2" />
-    <PackageReference Include="Swashbuckle.AspNetCore" Version="6.2.3" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="Dashboards\" />
-  </ItemGroup>
-
-</Project>
